feat: move start_device_tasks function to mqtt_handler for better organization and clarity
