+++ GreeMQTT/managers.py
@@ -3,31 +3,9 @@ import asyncio
 from aiomqtt import Client
 
 from GreeMQTT.logger import log
-from GreeMQTT.config import MQTT_QOS, MQTT_RETAIN
 from GreeMQTT.device.device import Device
 from GreeMQTT import device_db
-from GreeMQTT.mqtt_handler import get_params, subscribe
-
-
-async def start_device_tasks(
-    device: Device,
-    mqtt_client,
-    stop_event: asyncio.Event,
-):
-    """
-    Start async tasks for handling device parameters.
-    :param device:
-    :param mqtt_client:
-    :param stop_event:
-    :return:
-    """
-    asyncio.create_task(device.synchronize_time())
-
-    asyncio.create_task(
-        get_params(device, mqtt_client, stop_event, MQTT_QOS, MQTT_RETAIN)
-    )
-    asyncio.create_task(subscribe(device, mqtt_client, MQTT_QOS))
-    log.info("Started tasks for device", device=str(device))
+from GreeMQTT.mqtt_handler import start_device_tasks
 
 
 class DeviceRetryManager:
